Revert "Refactor: Use useGetItems hook for items in MeatballMenu component"
This reverts commit ee1f753bcdf8decadc13dae08024c4a95045b137.
